+++ Sources/PlayerKit/DesktopVLCPlayerWrapper.swift
@@ -700,6 +700,8 @@ extension DesktopVLCPlayerWrapper: StreamingInfoProtocol {
 
 extension DesktopVLCPlayerWrapper: PlayerEventSource {}
 
+extension DesktopVLCPlayerWrapper: PlayerMuteControlling {}
+
 extension DesktopVLCPlayerWrapper: PlayerPictureInPictureSupporting {
     var isPictureInPictureSupported: Bool { false }
     var isPictureInPicturePossible: Bool { false }